+++ packages/core/test/render3/es2015-tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "compilerOptions": {
+    "target": "es2015"
+  }
+}
